Move test_utils out of test folder so cz tests can use checkGradWrt
