+++ src/smith/differentiable_numerics/tests/CMakeLists.txt
@@ -4,16 +4,7 @@
 #
 # SPDX-License-Identifier: (BSD-3-Clause)
 
-set(differentiable_numerics_test_headers
-    differentiable_test_utils.hpp
-    )
-
-smith_add_library(
-    NAME        differentiable_numerics_test_utils
-    HEADERS     ${differentiable_numerics_test_headers}
-    )
-
-set(differentiable_numerics_test_depends smith_physics smith_differentiable_numerics differentiable_numerics_test_utils gtest)
+set(differentiable_numerics_test_depends smith_physics smith_differentiable_numerics gtest)
 
 set(differentiable_numerics_test_source
     test_field_state.cpp
